# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalyticsResponse": {
        "additionalProperties": false,
        "properties": {
          "failure_patterns": {
            "items": {
              "$ref": "#/components/schemas/FailurePattern"
            },
            "title": "Failure Patterns",
            "type": "array"
          },
          "rejected_count": {
            "minimum": 0.0,
            "title": "Rejected Count",
            "type": "integer"
          },
          "run_count": {
            "minimum": 0.0,
            "title": "Run Count",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "verified_count": {
            "minimum": 0.0,
            "title": "Verified Count",
            "type": "integer"
          }
        },
        "required": [
          "source",
          "run_count",
          "verified_count",
          "rejected_count",
          "failure_patterns"
        ],
        "title": "AnalyticsResponse",
        "type": "object"
      },
      "BaselineResult": {
        "additionalProperties": false,
        "properties": {
          "observed_status": {
            "anyOf": [
              {
                "maximum": 599.0,
                "minimum": 100.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Status"
          },
          "reproduced": {
            "title": "Reproduced",
            "type": "boolean"
          }
        },
        "required": [
          "reproduced"
        ],
        "title": "BaselineResult",
        "type": "object"
      },
      "ChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "max_challenges": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Challenges",
            "type": "integer"
          },
          "policy_ids": {
            "items": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            },
            "maxItems": 20,
            "title": "Policy Ids",
            "type": "array"
          }
        },
        "title": "ChallengeRequest",
        "type": "object"
      },
      "ChallengeResponse": {
        "additionalProperties": false,
        "properties": {
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "const": "execution",
            "default": "execution",
            "title": "Source",
            "type": "string"
          },
          "status": {
            "const": "challenging",
            "default": "challenging",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "ChallengeResponse",
        "type": "object"
      },
      "CreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "max_attempts": {
            "default": 3,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "type": "integer"
          },
          "target": {
            "const": "LedgerLite",
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "not_implemented",
              "not_found",
              "invalid_request",
              "conflict",
              "internal_error"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EventSeverity": {
        "enum": [
          "info",
          "warning",
          "error",
          "critical"
        ],
        "title": "EventSeverity",
        "type": "string"
      },
      "EventType": {
        "enum": [
          "stage_started",
          "stage_completed",
          "finding_discovered",
          "baseline_reproduced",
          "patch_proposed",
          "patch_applied",
          "test_completed",
          "challenge_proposed",
          "retry_scheduled",
          "run_completed",
          "run_failed"
        ],
        "title": "EventType",
        "type": "string"
      },
      "EventsResponse": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/SecurityEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          }
        },
        "required": [
          "run_id",
          "source",
          "events"
        ],
        "title": "EventsResponse",
        "type": "object"
      },
      "EvidenceReference": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Artifact Id",
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "sha256",
          "description"
        ],
        "title": "EvidenceReference",
        "type": "object"
      },
      "FailurePattern": {
        "additionalProperties": false,
        "properties": {
          "challenge_family": {
            "title": "Challenge Family",
            "type": "string"
          },
          "executions": {
            "minimum": 0.0,
            "title": "Executions",
            "type": "integer"
          },
          "failures": {
            "minimum": 0.0,
            "title": "Failures",
            "type": "integer"
          }
        },
        "required": [
          "challenge_family",
          "failures",
          "executions"
        ],
        "title": "FailurePattern",
        "type": "object"
      },
      "Finding": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Id",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/FindingSeverity"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "severity"
        ],
        "title": "Finding",
        "type": "object"
      },
      "FindingSeverity": {
        "enum": [
          "info",
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "FindingSeverity",
        "type": "string"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "properties": {
          "service": {
            "const": "proofloop",
            "default": "proofloop",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "HealthResponse",
        "type": "object"
      },
      "JsonValue": {},
      "PatchProposal": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "diff": {
            "maxLength": 200000,
            "minLength": 1,
            "title": "Diff",
            "type": "string"
          }
        },
        "required": [
          "attempt",
          "diff"
        ],
        "title": "PatchProposal",
        "type": "object"
      },
      "ReportResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "summary": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "source",
          "status",
          "summary",
          "limitations"
        ],
        "title": "ReportResponse",
        "type": "object"
      },
      "RunResponse": {
        "additionalProperties": false,
        "properties": {
          "baseline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BaselineResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/SecurityEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "finding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Finding"
              },
              {
                "type": "null"
              }
            ]
          },
          "patch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatchProposal"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "target": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Target",
            "type": "string"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "status",
          "target",
          "source"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "pending",
          "discovering",
          "reproducing",
          "generating_patch",
          "applying_patch",
          "verifying",
          "challenging",
          "retrying",
          "verified",
          "rejected",
          "inconclusive",
          "error"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "SecurityEvent": {
        "additionalProperties": false,
        "properties": {
          "event_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Metadata",
            "type": "object"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Run Id",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/EventSeverity"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "run_id",
          "timestamp",
          "stage",
          "event_type",
          "severity",
          "message",
          "source"
        ],
        "title": "SecurityEvent",
        "type": "object"
      },
      "Source": {
        "enum": [
          "fixture",
          "execution"
        ],
        "title": "Source",
        "type": "string"
      },
      "VerificationSummary": {
        "additionalProperties": false,
        "properties": {
          "adversarial_passed": {
            "minimum": 0.0,
            "title": "Adversarial Passed",
            "type": "integer"
          },
          "adversarial_total": {
            "minimum": 0.0,
            "title": "Adversarial Total",
            "type": "integer"
          },
          "functional_passed": {
            "minimum": 0.0,
            "title": "Functional Passed",
            "type": "integer"
          },
          "functional_total": {
            "minimum": 0.0,
            "title": "Functional Total",
            "type": "integer"
          },
          "security_passed": {
            "minimum": 0.0,
            "title": "Security Passed",
            "type": "integer"
          },
          "security_total": {
            "minimum": 0.0,
            "title": "Security Total",
            "type": "integer"
          }
        },
        "required": [
          "security_passed",
          "security_total",
          "functional_passed",
          "functional_total",
          "adversarial_passed",
          "adversarial_total"
        ],
        "title": "VerificationSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ProofLoop",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analytics": {
      "get": {
        "operationId": "analytics_api_analytics_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "pattern": "^[A-Za-z0-9_-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PLANNED: engine/integration not implemented"
          }
        },
        "summary": "Analytics"
      }
    },
    "/api/health": {
      "get": {
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/runs": {
      "post": {
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PLANNED: engine/integration not implemented"
          }
        },
        "summary": "Create Run"
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PLANNED: engine/integration not implemented"
          }
        },
        "summary": "Get Run"
      }
    },
    "/api/runs/{run_id}/challenge": {
      "post": {
        "operationId": "challenge_api_runs__run_id__challenge_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PLANNED: engine/integration not implemented"
          }
        },
        "summary": "Challenge"
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "operationId": "get_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PLANNED: engine/integration not implemented"
          }
        },
        "summary": "Get Events"
      }
    },
    "/api/runs/{run_id}/report": {
      "get": {
        "operationId": "get_report_api_runs__run_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PLANNED: engine/integration not implemented"
          }
        },
        "summary": "Get Report"
      }
    }
  }
}